Refactoring automatic assembly logic
