+++ validator-lite-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,7 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.viiyue.validator.spring.boot.autoconfigure.configuration.RemoveDefaultValidationBeanAutoConfiguration,\
+com.viiyue.validator.spring.boot.autoconfigure.configuration.ValidatorInitialAutoConfiguration,\
+com.viiyue.validator.spring.boot.autoconfigure.configuration.RedefineRequestMappingHandlerAdapterAutoConfiguration,\
 com.viiyue.validator.spring.boot.autoconfigure.configuration.ValidationAutoConfiguration
 
 # Application Listeners
